scripts/qwen-vl-finetune/requirements.txt
# PTCG Qwen-VL Fine-tuning Dependencies
# Optimized for RTX 5070 Ti 16GB VRAM
# Python 3.10+ required

# =============================================================================
# Core ML Libraries
# =============================================================================
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Install with CUDA 12.1 support:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# =============================================================================
# Transformers & Training
# =============================================================================
transformers>=4.37.0
datasets>=2.14.0
trl>=0.7.10
accelerate>=0.25.0
peft>=0.7.0

# =============================================================================
# Quantization (QLoRA 4-bit)
# =============================================================================
bitsandbytes>=0.41.0

# =============================================================================
# Qwen VL Utilities
# =============================================================================
qwen-vl-utils>=0.0.1

# =============================================================================
# Environment & Database
# =============================================================================
python-dotenv>=1.0.0
prisma>=0.11.0

# =============================================================================
# Image Processing (Memory Optimized)
# =============================================================================
Pillow>=10.0.0
opencv-python>=4.8.0

# =============================================================================
# Utilities
# =============================================================================
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.0.0

# =============================================================================
# Evaluation
# =============================================================================
scikit-learn>=1.3.0
jiwer>=3.0.0

# =============================================================================
# Logging & Visualization
# =============================================================================
tensorboard>=2.15.0
matplotlib>=3.8.0
seaborn>=0.13.0

# =============================================================================
# API (for Inference Service)
# =============================================================================
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0

# =============================================================================
# Optional: Flash Attention 2 (Speeds up training)
# =============================================================================
# pip install flash-attn --no-build-isolation
# Note: Requires Ampere or newer GPU (RTX 30xx, 40xx, 50xx series)
